+++ test/bones_api_test.reflection.g.dart
@@ -1,6 +1,6 @@
 //
 // GENERATED CODE - DO NOT MODIFY BY HAND!
-// BUILDER: reflection_factory/1.2.12
+// BUILDER: reflection_factory/1.2.13
 // BUILD COMMAND: dart run build_runner build
 //
 
@@ -36,7 +36,7 @@ class MyInfoModule$reflection extends ClassReflection<MyInfoModule> {
   Version get languageVersion => Version.parse('2.17.0');
 
   @override
-  Version get reflectionFactoryVersion => Version.parse('1.2.12');
+  Version get reflectionFactoryVersion => Version.parse('1.2.13');
 
   @override
   MyInfoModule$reflection withObject([MyInfoModule? obj]) =>